+++ viadot/sources/sharepoint.py
@@ -10,7 +10,7 @@ from viadot.exceptions import CredentialError
 
 from ..config import get_source_credentials
 from ..signals import SKIP
-from ..utils import add_viadot_metadata_columns, cleanup_df, validations
+from ..utils import add_viadot_metadata_columns, cleanup_df, validate
 from .base import Source
 
 
@@ -113,7 +113,7 @@ class Sharepoint(Source):
                 Specify None to get all worksheets. Defaults to None.
             if_empty (str, optional): What to do if the file is empty. Defaults to "warn".
             tests (Dict[str], optional): A dictionary with optional list of tests
-                to verify the output dataframe. If defined, triggers the `validations`
+                to verify the output dataframe. If defined, triggers the `validate`
                 function from utils. Defaults to None.
             kwargs (dict[str, Any], optional): Keyword arguments to pass to pd.ExcelFile.parse(). Note that
             `nrows` is not supported.
@@ -158,6 +158,6 @@ class Sharepoint(Source):
         df_clean = cleanup_df(df)
 
         if tests:
-            validations(df=df_clean, tests=tests)
+            validate(df=df_clean, tests=tests)
 
         return df_clean